refactor(sidecar): own resource lifecycle and browser snapshots
