+++ src/react-workbench/chat/ChatPage.tsx
@@ -1,5 +1,6 @@
+import { SidecarResources, initialSidecarLayout, type SidecarResourcesHandle, type SidecarLayout } from "../sidecar/SidecarResources";
 import { useChatSubmission } from "./useChatSubmission";
-import { lazy, Suspense, useCallback, useEffect, useEffectEvent, useLayoutEffect, useMemo, useReducer, useRef, useState, type CSSProperties, type KeyboardEvent as ReactKeyboardEvent } from "react";
+import { useCallback, useEffect, useEffectEvent, useLayoutEffect, useMemo, useReducer, useRef, useState, type CSSProperties, type KeyboardEvent as ReactKeyboardEvent } from "react";
 import { elementTransitions, useExitPresence } from "../lib/useExitPresence";
 import type { TFunction } from "i18next";
 import {
@@ -43,7 +44,6 @@ import { reduceSessionDeleteState } from "../sessions/sessionDeleteState";
 import type { ToolCallSummary } from "./messageActions";
 import type { AgentUiForm } from "../../app-core/agent-ui/agentUiEvents";
 import { AgentUiFormCard } from "./AgentUiFormCard";
-import { DataViewCard } from "./DataViewCard";
 import {
   projectChatEventEffects,
   projectTimelineSessionStatus,
@@ -64,31 +64,20 @@ import {
 } from "./sessionTabWorkspace";
 import {
   groupSessionsByWorkspace,
-  sessionWorkspaceName,
 } from "./sessionWorkspaces";
 import {
   applyLoadedDelegatedAgentTrace,
-  projectLoadedArtifactDetail,
 } from "../../app-core/chat/chatProjection";
 import type {
   ArtifactRef,
   DelegatedAgentState,
-  LoadedArtifactDetail,
 } from "../../app-core/chat/chatTurnContracts";
 import {
-  type OfficeArtifactSource,
   type SpreadsheetCellChangeRequest,
 } from "../../app-core/chat/officeArtifact";
-import { officeContentReference } from "../../app-core/chat/officeContentReference";
-import { ArtifactReviewPanel } from "../sidecar/ArtifactReviewPanel";
-import { useArtifactFile } from "./useArtifactFile";
 import type { AgentInputReference } from "../../app-core/chat/agentInputReference";
 import { logRendererEvent } from "../../app-core/native/rendererLogger";
 import type { ChatTimelineSnapshot } from "../../app-core/chat/agentTimelineModel";
-import type {
-  NativeBrowserSession,
-  NativeBrowserSnapshot,
-} from "../../app-core/native/nativeBrowserSnapshot";
 import {
   isThreadCommandInFlight,
   type ThreadCommandLifecycle,
@@ -110,14 +99,7 @@ import { ChatTimeline } from "./ChatTimeline";
 import { captureConversationView, restoreConversationView, type ConversationViewState } from "./conversationViewport";
 import { EmptyChatStart } from "./EmptyChatStart";
 import { FloatingPlanStatus } from "./FloatingPlanStatus";
-import { AssistantMarkdown } from "./AssistantMarkdown";
-import {
-  AssistantFileLinkError,
-  assistantFileArtifact,
-  assistantFileLinkTitle,
-  resolveAssistantFileLink,
-  type AssistantFileLink,
-} from "./assistantFileLinks";
+import type { AssistantFileLink } from "./assistantFileLinks";
 import {
   ChatSessionWorkspace,
   type ProjectSessionContext,
@@ -127,23 +109,6 @@ import {
   isDefaultSessionTitle,
 } from "./sessionTitle";
 import { projectTinybotMascotMood, type TinybotMascotMood } from "./TinybotMascot";
-import { Sidecar } from "../sidecar/Sidecar";
-import { SidecarBrowser } from "../sidecar/SidecarBrowser";
-import { OfficeArtifactPreview } from "../sidecar/OfficeArtifactPreview";
-import {
-  activeSidecarTab,
-  createInitialSidecarState,
-  DEFAULT_SIDECAR_WORKSPACE_ID,
-  readPersistedSidecarWidth,
-  reduceSidecarState,
-  sidecarArtifactTabId,
-  visibleSidecarTabs,
-  writePersistedSidecarWidth,
-  type SidecarArtifactTab,
-  type SidecarBrowserTab,
-  type SidecarTab,
-  type SidecarTerminalTab,
-} from "../sidecar/sidecarModel";
 
 export type ChatPageProps = {
   chatStore: ChatStore;
@@ -182,23 +147,6 @@ type DrawerState =
   | { kind: "tool"; title: string; toolCall: ToolCallSummary }
   | { kind: "subagent"; title: string; delegate: DelegatedAgentState; loading: boolean; error?: string }
   | null;
-
-type ArtifactSidecarContent = {
-  localFile?: boolean;
-  artifact: ArtifactRef;
-  detail?: LoadedArtifactDetail;
-  error?: string;
-  loading: boolean;
-  notice?: string;
-  office?: OfficeArtifactSource;
-};
-
-type BrowserSnapshot = NativeBrowserSnapshot<NativeBrowserSession>;
-
-const LazySidecarTerminal = lazy(async () => {
-  const module = await import("../sidecar/SidecarTerminal");
-  return { default: module.SidecarTerminal };
-});
 
 function resolveComposerModel(
   models: readonly ModelOption[],
@@ -353,16 +301,9 @@ export function ChatPage({
   const drawerElementRef = useRef<HTMLElement>(null);
   const drawerTriggerRef = useRef<HTMLElement | null>(null);
   const sidecarToggleRef = useRef<HTMLButtonElement>(null);
-  const [sidecar, dispatchSidecar] = useReducer(
-    reduceSidecarState,
-    undefined,
-    () => createInitialSidecarState(readPersistedSidecarWidth(window.localStorage)),
-  );
-  const [artifactSidecarContent, setArtifactSidecarContent] = useState<Record<string, ArtifactSidecarContent>>({});
+  const sidecarResources = useRef<SidecarResourcesHandle>(null);
+  const [sidecar, setSidecar] = useState<SidecarLayout>(initialSidecarLayout);
   const [composerFocusRequestId, setComposerFocusRequestId] = useState(0);
-  const [browserProvisionErrors, setBrowserProvisionErrors] = useState<Record<string, string>>({});
-  const [terminalErrors, setTerminalErrors] = useState<Record<string, string>>({});
-  const [browserProvisionEpoch, setBrowserProvisionEpoch] = useState(0);
   const [composerSessionMentionIds, setComposerSessionMentionIds] = useState<string[]>([]);
   const [composerSelectedSkillIds, setComposerSelectedSkillIds] = useState<string[]>([]);
   const [composerArtifactReferences, setComposerArtifactReferences] = useState<(AgentInputReference & { id: string })[]>([]);
@@ -387,10 +328,6 @@ export function ChatPage({
   const pendingConversationRestoreRef = useRef("");
   const hasActivatedSessionRef = useRef(false);
   const stickToLatestRef = useRef(true);
-  const sidecarRef = useRef(sidecar);
-  const browserProvisioningResourceIdRef = useRef("");
-  const browserActivationTargetRef = useRef("");
-  sidecarRef.current = sidecar;
   sessionTabsRef.current = sessionTabs;
   sessionsLoadedRef.current = sessionsLoaded;
   const activeSessionId = sessionTabs.activeSessionId;
@@ -429,16 +366,11 @@ export function ChatPage({
   });
   const {
     agentUiForms,
-    browserError,
-    browserSnapshot,
     error: timelineError,
     hookResults,
     timeline,
   } = sessionRuntime.state;
   const {
-    acceptBrowserSnapshot,
-    clearBrowserError,
-    clearBrowserSnapshot,
     clearError: clearTimelineError,
     reload: reloadSessionRuntime,
     reportError: reportTimelineError,
@@ -483,16 +415,6 @@ export function ChatPage({
       label: annotation.fileTitle,
     }))]
   ), [composerArtifactReferences, composerSpreadsheetAnnotations, t]);
-  const sidecarTabs = useMemo(() => visibleSidecarTabs(sidecar), [sidecar]);
-  const sidecarActiveTab = useMemo(() => activeSidecarTab(sidecar), [sidecar]);
-  const explicitWorkspaceId = activeDisplaySession?.workingDirectory?.trim() ?? "";
-  const activeWorkspaceId = activeDisplaySession
-    ? explicitWorkspaceId || DEFAULT_SIDECAR_WORKSPACE_ID
-    : "";
-  const activeWorkspaceLabel = explicitWorkspaceId
-    ? sessionWorkspaceName(explicitWorkspaceId)
-    : activeDisplaySession ? t("shell.generalSessions") : "";
-
   useEffect(() => {
     if (!toolsStore?.loadCatalog) {
       setComposerSkills([]);
@@ -523,166 +445,6 @@ export function ChatPage({
       cancelled = true;
     };
   }, [activeDisplaySession?.pluginMigration, activeDisplaySession?.workingDirectory, toolsStore]);
-  const unboundBrowserResource = useMemo(() => sidecar.tabs.find((tab): tab is SidecarBrowserTab => (
-    tab.kind === "browser"
-      && tab.threadId === activeSession?.id
-      && !tab.nativeTabId
-  )), [activeSession?.id, sidecar.tabs]);
-  const retainedBrowserResource = useMemo(() => sidecar.tabs.find((tab): tab is SidecarBrowserTab => (
-    tab.kind === "browser"
-      && tab.threadId === activeSession?.id
-      && Boolean(tab.browserSessionId)
-      && Boolean(tab.nativeTabId)
-  )), [activeSession?.id, sidecar.tabs]);
-
-  const synchronizeBrowserSnapshot = useCallback((snapshot: BrowserSnapshot, acceptForActiveThread = true) => {
-    if (acceptForActiveThread && snapshot.data.sessionId === activeSessionId) {
-      acceptBrowserSnapshot(snapshot);
-    }
-    dispatchSidecar({
-      browserSessionId: snapshot.data.browserSessionId,
-      tabs: snapshot.data.tabs.map((tab) => ({
-        nativeTabId: tab.tabId,
-        title: browserResourceTitle(tab.title, tab.url, t("sidecar.browser")),
-      })),
-      threadId: snapshot.data.sessionId,
-      type: "tab.syncBrowserSession",
-    });
-  }, [acceptBrowserSnapshot, activeSessionId, t]);
-
-  useEffect(() => {
-    dispatchSidecar({
-      threadId: activeSession?.id ?? "",
-      type: "scope.changed",
-      workspaceId: activeWorkspaceId,
-    });
-  }, [activeSession?.id, activeWorkspaceId]);
-
-  useEffect(() => {
-    if (browserSnapshot) synchronizeBrowserSnapshot(browserSnapshot, false);
-  }, [browserSnapshot, synchronizeBrowserSnapshot]);
-
-  useEffect(() => {
-    const resource = retainedBrowserResource;
-    const browserRuntime = chatStore.browserRuntime;
-    if (!resource?.browserSessionId
-      || !browserRuntime
-      || browserSnapshot?.data.browserSessionId === resource.browserSessionId) return;
-    let cancelled = false;
-    void browserRuntime.snapshot(resource.browserSessionId)
-      .then((snapshot) => {
-        if (cancelled) return;
-        if (snapshot.data.sessionId !== resource.threadId) {
-          throw new Error(
-            `Browser snapshot session ${snapshot.data.sessionId} does not match resource thread ${resource.threadId}.`,
-          );
-        }
-        synchronizeBrowserSnapshot(snapshot);
-      })
-      .catch((error) => {
-        if (!cancelled) {
-          setBrowserProvisionErrors((current) => ({ ...current, [resource.id]: errorMessage(error) }));
-        }
-      });
-    return () => {
-      cancelled = true;
-    };
-  }, [
-    browserProvisionEpoch,
-    browserSnapshot?.data.browserSessionId,
-    chatStore.browserRuntime,
-    retainedBrowserResource,
-    synchronizeBrowserSnapshot,
-  ]);
-
-  useEffect(() => {
-    const resource = unboundBrowserResource;
-    const browserRuntime = chatStore.browserRuntime;
-    if (!resource
-      || browserProvisionErrors[resource.id]
-      || browserProvisioningResourceIdRef.current) return;
-    browserProvisioningResourceIdRef.current = resource.id;
-    void (async () => {
-      try {
-        if (!browserRuntime) throw new Error(t("sidecar.browserBuildUnavailable"));
-        let snapshot = await browserRuntime.createSession({ ownerSessionId: resource.threadId });
-
-        const currentResources = sidecarRef.current.tabs.filter((tab): tab is SidecarBrowserTab => (
-          tab.kind === "browser" && tab.threadId === resource.threadId
-        ));
-        const currentResource = currentResources.find((tab) => tab.id === resource.id);
-        const resourceStillExists = Boolean(currentResource);
-        const resourceAlreadyBound = Boolean(
-          currentResource?.browserSessionId === snapshot.data.browserSessionId
-            && currentResource.nativeTabId
-            && snapshot.data.tabs.some((tab) => tab.tabId === currentResource.nativeTabId),
-        );
-        const boundNativeTabIds = new Set(currentResources.flatMap((tab) => tab.nativeTabId ? [tab.nativeTabId] : []));
-        const hasUnboundNativeTab = snapshot.data.tabs.some((tab) => !boundNativeTabIds.has(tab.tabId));
-        let createdNativeTabId = "";
-        if (resourceStillExists && !resourceAlreadyBound && !hasUnboundNativeTab) {
-          const previousNativeTabIds = new Set(snapshot.data.tabs.map((tab) => tab.tabId));
-          snapshot = await browserRuntime.createTab(snapshot.data.browserSessionId);
-          createdNativeTabId = snapshot.data.tabs.find((tab) => !previousNativeTabIds.has(tab.tabId))?.tabId ?? "";
-        }
-
-        if (!sidecarRef.current.tabs.some((tab) => tab.id === resource.id)) {
-          if (createdNativeTabId && snapshot.data.tabs.length > 1) {
-            await browserRuntime.closeTab(snapshot.data.browserSessionId, createdNativeTabId);
-          }
-          return;
-        }
-        synchronizeBrowserSnapshot(snapshot);
-      } catch (error) {
-        if (sidecarRef.current.tabs.some((tab) => tab.id === resource.id)) {
-          setBrowserProvisionErrors((current) => ({ ...current, [resource.id]: errorMessage(error) }));
-        }
-      } finally {
-        if (browserProvisioningResourceIdRef.current === resource.id) {
-          browserProvisioningResourceIdRef.current = "";
-        }
-        setBrowserProvisionEpoch((current) => current + 1);
-      }
-    })();
-  }, [
-    browserProvisionEpoch,
-    browserProvisionErrors,
-    chatStore.browserRuntime,
-    synchronizeBrowserSnapshot,
-    t,
-    unboundBrowserResource,
-  ]);
-
-  useEffect(() => {
-    const resource = sidecarActiveTab?.kind === "browser" ? sidecarActiveTab : undefined;
-    const browserRuntime = chatStore.browserRuntime;
-    if (!resource?.browserSessionId
-      || !resource.nativeTabId
-      || !browserRuntime
-      || browserSnapshot?.data.browserSessionId !== resource.browserSessionId) return;
-    const activationTarget = `${resource.browserSessionId}:${resource.nativeTabId}`;
-    if (browserSnapshot.data.activeTabId === resource.nativeTabId) {
-      if (browserActivationTargetRef.current === activationTarget) {
-        browserActivationTargetRef.current = "";
-      }
-      return;
-    }
-    if (browserActivationTargetRef.current === activationTarget) return;
-    browserActivationTargetRef.current = activationTarget;
-    void browserRuntime.activateTab(resource.browserSessionId, resource.nativeTabId)
-      .then((snapshot) => synchronizeBrowserSnapshot(snapshot))
-      .catch((error) => {
-        if (browserActivationTargetRef.current === activationTarget) {
-          browserActivationTargetRef.current = "";
-        }
-        setBrowserProvisionErrors((current) => ({ ...current, [resource.id]: errorMessage(error) }));
-      });
-  }, [browserSnapshot, chatStore.browserRuntime, sidecarActiveTab, synchronizeBrowserSnapshot]);
-
-  useEffect(() => {
-    writePersistedSidecarWidth(window.localStorage, sidecar.width);
-  }, [sidecar.width]);
-
   useEffect(() => {
     setMigrationInstallError("");
   }, [activeSessionId]);
@@ -1434,229 +1196,8 @@ export function ChatPage({
     }
   }
 
-  async function handleOpenArtifact(artifact: ArtifactRef) {
-    if (!activeSession) {
-      return;
-    }
-    const tabId = sidecarArtifactTabId(activeSession.id, artifact.id);
-    dispatchSidecar({
-      artifactId: artifact.id,
-      threadId: activeSession.id,
-      title: artifact.title,
-      type: "tab.openArtifact",
-    });
-    if (artifact.kind === "data_view") {
-      setArtifactSidecarContent((current) => ({
-        ...current,
-        [tabId]: {
-          artifact,
-          ...(artifact.dataView ? { detail: { id: artifact.id, title: artifact.title, mimeType: artifact.mimeType, dataView: artifact.dataView } } : {}),
-          loading: false,
-          ...(artifact.dataViewError ? { error: artifact.dataViewError } : {}),
-        },
-      }));
-      return;
-    }
-    setArtifactSidecarContent((current) => ({
-      ...current,
-      [tabId]: { artifact, loading: Boolean(chatStore.loadArtifact) },
-    }));
-    if (!chatStore.loadArtifact) {
-      return;
-    }
-    try {
-      const payload = await chatStore.loadArtifact({
-        artifactId: artifact.id,
-        sessionKey: activeSession.id,
-      });
-      const detail = projectLoadedArtifactDetail(artifact, payload);
-      setArtifactSidecarContent((current) => current[tabId]
-        ? { ...current, [tabId]: { ...current[tabId], detail, loading: false } }
-        : current);
-    } catch (error) {
-      const message = error instanceof Error ? error.message : String(error);
-      setArtifactSidecarContent((current) => current[tabId]
-        ? { ...current, [tabId]: { ...current[tabId], error: message, loading: false } }
-        : current);
-    }
-  }
-
-  async function handleOpenAssistantFileLink(link: AssistantFileLink) {
-    if (!activeSession) {
-      return;
-    }
-
-    let artifact: ArtifactRef;
-    try {
-      artifact = assistantFileArtifact(resolveAssistantFileLink(link.href, activeSession.workingDirectory));
-      logRendererEvent("info", "artifact.file_link.resolved", {
-        href: link.href, path: artifact.fetchPath, sessionId: activeSession.id,
-      });
-    } catch (error) {
-      artifact = assistantFileArtifact({ path: link.href, title: assistantFileLinkTitle(link.href) });
-      const tabId = sidecarArtifactTabId(activeSession.id, artifact.id);
-      dispatchSidecar({
-        artifactId: artifact.id,
-        threadId: activeSession.id,
-        title: artifact.title,
-        type: "tab.openArtifact",
-      });
-      const message = error instanceof AssistantFileLinkError && error.code === "outside_workspace"
-        ? t("details.fileOutsideWorkspace")
-        : errorMessage(error);
-      console.error("[artifact-preview] workspace file link resolution failed", {
-        error,
-        href: link.href,
-        sessionId: activeSession.id,
-        workspaceRoot: activeSession.workingDirectory,
-      });
-      setArtifactSidecarContent((current) => ({
-        ...current,
-        [tabId]: { artifact, error: message, loading: false },
-      }));
-      return;
-    }
-
-    const tabId = sidecarArtifactTabId(activeSession.id, artifact.id);
-    dispatchSidecar({
-      artifactId: artifact.id,
-      threadId: activeSession.id,
-      title: artifact.title,
-      type: "tab.openArtifact",
-    });
-    setArtifactSidecarContent((current) => ({
-      ...current,
-      [tabId]: { artifact, localFile: true, loading: false },
-    }));
-  }
-
-  async function handleCloseSidecarTab(tab: SidecarTab) {
-    if (tab.kind === "browser") {
-      setBrowserProvisionErrors((current) => omitRecordKey(current, tab.id));
-      const browserRuntime = chatStore.browserRuntime;
-      if (!browserRuntime || !tab.browserSessionId || !tab.nativeTabId) {
-        dispatchSidecar({ tabId: tab.id, type: "tab.close" });
-        return;
-      }
-      try {
-        let snapshot = await browserRuntime.snapshot(tab.browserSessionId);
-        const remainingResources = sidecarRef.current.tabs.filter((candidate) => (
-          candidate.kind === "browser"
-            && candidate.threadId === tab.threadId
-            && candidate.id !== tab.id
-        ));
-        if (snapshot.data.tabs.length === 1 && remainingResources.length) {
-          snapshot = await browserRuntime.createTab(snapshot.data.browserSessionId);
-        }
-        if (snapshot.data.tabs.length === 1) {
-          await browserRuntime.closeSession(snapshot.data.browserSessionId);
-          clearBrowserSnapshot(snapshot.data.browserSessionId);
-          dispatchSidecar({ tabId: tab.id, type: "tab.close" });
-          return;
-        }
-        const next = await browserRuntime.closeTab(snapshot.data.browserSessionId, tab.nativeTabId);
-        dispatchSidecar({ tabId: tab.id, type: "tab.close" });
-        synchronizeBrowserSnapshot(next);
-      } catch (error) {
-        setBrowserProvisionErrors((current) => ({ ...current, [tab.id]: errorMessage(error) }));
-      }
-      return;
-    }
-
-    if (tab.kind === "terminal") {
-      setTerminalErrors((current) => omitRecordKey(current, tab.id));
-      try {
-        await chatStore.terminalRuntime?.terminate(tab.id);
-      } catch (error) {
-        setTerminalErrors((current) => ({ ...current, [tab.id]: errorMessage(error) }));
-        return;
-      }
-      dispatchSidecar({ tabId: tab.id, type: "tab.close" });
-      return;
-    }
-
-    dispatchSidecar({ tabId: tab.id, type: "tab.close" });
-    if (tab.kind === "artifact") {
-      setArtifactSidecarContent((current) => omitRecordKey(current, tab.id));
-    }
-  }
-
-  function renderSidecarArtifact(tab: SidecarArtifactTab) {
-    const content = artifactSidecarContent[tab.id];
-    if (!content) {
-      return <p className="react-empty-state">{t("details.noPreview")}</p>;
-    }
-    return (
-      <ArtifactDetails
-        key={tab.id}
-        reviewEpoch={artifactReviewEpoch}
-        responding={sessionResponding}
-        localThreadId={content.localFile ? tab.threadId : undefined}
-        observeFile={sidecar.presentation !== "closed" && tab.threadId === activeSessionId}
-        workspaceStore={workspaceStore}
-        onReference={(reference) => {
-          const id = "artifact:" + tab.id + ":" + reference.detail;
-          setComposerArtifactReferences((current) => [...current.filter((item) => item.id !== id), { ...reference, id }]);
-          setComposerFocusRequestId((current) => current + 1);
-        }}
-        artifact={content.artifact}
-        detail={content.detail}
-        error={content.error}
-        loading={content.loading}
-        notice={content.notice}
-        office={content.office}
-        onAskForSpreadsheetChange={handleSpreadsheetAskForChange}
-        onOpenFileLink={handleOpenAssistantFileLink}
-      />
-    );
-  }
-
-  function renderSidecarBrowser(tab: SidecarBrowserTab, surfaceVisible: boolean) {
-    return (
-      <SidecarBrowser
-        browserRuntime={chatStore.browserRuntime}
-        externalError={browserProvisionErrors[tab.id] || browserError}
-        snapshot={browserSnapshot?.data.sessionId === tab.threadId ? browserSnapshot : undefined}
-        surfaceVisible={surfaceVisible && !presentDrawer}
-        tab={tab}
-        onHandoffComplete={() => handleBrowserHandoffComplete(tab)}
-        onRetryProvision={() => {
-          clearBrowserError();
-          setBrowserProvisionErrors((current) => omitRecordKey(current, tab.id));
-          setBrowserProvisionEpoch((current) => current + 1);
-        }}
-        onSnapshot={synchronizeBrowserSnapshot}
-      />
-    );
-  }
-
-  function renderSidecarTerminal(tab: SidecarTerminalTab) {
-    return (
-      <Suspense fallback={(
-        <div aria-busy="true" className="react-sidecar__deferred" role="status">
-          <Loader2 aria-hidden="true" size={18} />
-          <span>{t("sidecar.terminalStarting")}</span>
-        </div>
-      )}>
-        <LazySidecarTerminal
-          externalError={terminalErrors[tab.id]}
-          tab={tab}
-          terminalRuntime={chatStore.terminalRuntime}
-          workspaceLabel={activeWorkspaceLabel}
-        />
-      </Suspense>
-    );
-  }
-
-  async function handleBrowserHandoffComplete(tab: SidecarBrowserTab) {
-    if (!activeSession || activeSession.id !== tab.threadId) return;
-    try {
-      await submission.submitTurn(activeSession.id, { text: t("browserHandoffContinue") }, "browser-handoff-complete");
-      await handleSessionStoreRefresh(activeSession);
-    } catch (error) {
-      reportTimelineError(t("sidecar.browserHandoffFailed", { message: errorMessage(error) }));
-    }
-  }
+  function handleOpenArtifact(artifact: ArtifactRef) { return sidecarResources.current?.openArtifact(artifact); }
+  function handleOpenAssistantFileLink(link: AssistantFileLink) { return sidecarResources.current?.openFileLink(link); }
 
   async function handleSubmitAgentUiForm(form: AgentUiForm, values: Record<string, unknown>) {
     await chatApplication.submitForm(activePersistedSessionId, form, values);
@@ -1704,11 +1245,6 @@ export function ChatPage({
       else setComposerFocusRequestId((current) => current + 1);
     }
     setDrawer(null);
-  }
-
-  function hideSidecar() {
-    sidecarToggleRef.current?.focus();
-    dispatchSidecar({ type: "presentation.hide" });
   }
 
   function handleComposerDraftChange(value: string) {
@@ -1813,9 +1349,7 @@ export function ChatPage({
               ref={sidecarToggleRef}
               title={sidecar.presentation === "closed" ? t("sidecar.show") : t("sidecar.hide")}
               type="button"
-              onClick={() => dispatchSidecar({
-                type: sidecar.presentation === "closed" ? "presentation.show" : "presentation.hide",
-              })}
+              onClick={() => sidecarResources.current?.toggle()}
             >
               {sidecar.presentation === "closed"
                 ? <PanelRightOpen aria-hidden="true" size={17} />
@@ -2050,24 +1584,28 @@ export function ChatPage({
         </div>
       </main>
 
-      <Sidecar
-        scopeKey={JSON.stringify([activeSessionId, activeWorkspaceId])}
-        activeTabId={sidecarActiveTab?.id ?? ""}
-        canCreateBrowser={Boolean(activeSession)}
-        canCreateTerminal={Boolean(activeWorkspaceId)}
-        presentation={sidecar.presentation}
-        renderArtifact={renderSidecarArtifact}
-        renderBrowser={renderSidecarBrowser}
-        renderTerminal={renderSidecarTerminal}
-        tabs={sidecarTabs}
-        width={sidecar.width}
-        onActivateTab={(tabId) => dispatchSidecar({ tabId, type: "tab.activate" })}
-        onCloseTab={handleCloseSidecarTab}
-        onCreateBrowser={() => dispatchSidecar({ type: "tab.newBrowser" })}
-        onCreateTerminal={(shell) => dispatchSidecar({ shell, type: "tab.newTerminal" })}
-        onHide={hideSidecar}
-        onResize={(width, maxWidth) => dispatchSidecar({ maxWidth, type: "presentation.resize", width })}
-        onToggleExpanded={() => dispatchSidecar({ type: "presentation.toggleExpanded" })}
+      <SidecarResources
+        ref={sidecarResources}
+        activeSession={activeSession}
+        activeDisplaySession={activeDisplaySession}
+        activeSessionId={activeSessionId}
+        chatStore={chatStore}
+        workspaceStore={workspaceStore}
+        artifactReviewEpoch={artifactReviewEpoch}
+        sessionResponding={sessionResponding}
+        presentDrawer={Boolean(presentDrawer)}
+        onLayoutChange={setSidecar}
+        onHide={() => sidecarToggleRef.current?.focus()}
+        onReference={(reference) => {
+          setComposerArtifactReferences((current) => [...current.filter((item) => item.id !== reference.id), reference]);
+          setComposerFocusRequestId((current) => current + 1);
+        }}
+        onAskForSpreadsheetChange={handleSpreadsheetAskForChange}
+        onHandoff={async (sessionId) => {
+          await submission.submitTurn(sessionId, { text: t("browserHandoffContinue") }, "browser-handoff-complete");
+          await handleSessionStoreRefresh(activeSession);
+        }}
+        onError={reportTimelineError}
       />
 
       {presentDrawer ? (
@@ -2223,116 +1761,6 @@ function SubagentDetails({
   );
 }
 
-function ArtifactDetails({
-  artifact,
-  detail: storedDetail,
-  error: storedError,
-  loading: storedLoading,
-  notice: storedNotice,
-  office: storedOffice,
-  localThreadId,
-  reviewEpoch,
-  responding,
-  observeFile,
-  workspaceStore,
-  onReference,
-  onAskForSpreadsheetChange,
-  onOpenFileLink,
-}: {
-  artifact: ArtifactRef;
-  detail?: LoadedArtifactDetail;
-  error?: string;
-  loading: boolean;
-  notice?: string;
-  office?: OfficeArtifactSource;
-  localThreadId?: string;
-  reviewEpoch: number;
-  responding: boolean;
-  observeFile: boolean;
-  workspaceStore?: Pick<WorkspaceStore, "readThreadFile" | "readThreadFileBytes" | "artifactReviews">;
-  onReference: (reference: AgentInputReference) => void;
-  onAskForSpreadsheetChange: (artifact: ArtifactRef, request: SpreadsheetCellChangeRequest, revision?: string) => void;
-  onOpenFileLink: (link: AssistantFileLink) => void;
-}) {
-  const { t } = useTranslation("chat");
-  const [refreshKey, setRefreshKey] = useState(0);
-  const file = useArtifactFile({
-    refreshKey,
-    artifact, enabled: Boolean(localThreadId) && observeFile, threadId: localThreadId, workspaceStore,
-    unavailableMessage: t("details.filePreviewUnavailable"), binaryMessage: t("details.binaryFilePreviewUnsupported"),
-  });
-  const { detail, error, loading, office } = localThreadId ? file : { detail: storedDetail, error: storedError, loading: storedLoading, office: storedOffice };
-  const notice = localThreadId ? (file.truncated ? t("details.filePreviewTruncated") : undefined) : storedNotice;
-  function referenceArtifact() {
-    const text = detail?.dataView ? JSON.stringify(detail.dataView) : detail?.textContent;
-    const excerpt = text && text.length > 12000 ? text.slice(0, 12000) + "\n[Preview excerpt truncated]" : text;
-    onReference({
-      kind: "reference", title: artifact.title, detail: t("details.artifactReference"),
-      ...(localThreadId ? { referenceKind: "file", sourcePath: artifact.fetchPath, scope: localThreadId, revision: file.revision } as const : {}),
-      sourceText: [
-        `Artifact: ${artifact.title}`, `Artifact ID: ${artifact.id}`,
-        ...(localThreadId ? [`File: ${artifact.fetchPath}`, `Viewed revision: ${file.revision}`, "Verify the current file before editing; this reference describes the viewed revision."] : []),
-        ...(excerpt ? [excerpt] : []),
-      ].join("\n"),
-    });
-  }
-  const markdown = isMarkdownArtifact(artifact, detail);
-  const markdownContent = detail?.textContent && markdown
-    ? { text: detail.textContent, title: detail.title }
-    : undefined;
-  return (
-    <div className="react-artifact-detail" data-content={markdown || office?.kind === "document" ? "document" : "preview"}>
-      <div className="react-artifact-detail__toolbar">
-        <button disabled={loading || Boolean(error)} onClick={referenceArtifact} type="button">{t("details.referenceInChat")}</button>
-        {localThreadId ? <span role="status">{t("details.fileAutoUpdates")}</span> : null}
-      </div>
-      {localThreadId && artifact.fetchPath && workspaceStore?.artifactReviews ? (
-        <ArtifactReviewPanel store={workspaceStore.artifactReviews} path={artifact.fetchPath} threadId={localThreadId}
-          revision={error ? undefined : file.revision} epoch={reviewEpoch} responding={responding}
-          kind={office?.kind} title={artifact.title} onRestored={() => setRefreshKey((value) => value + 1)} />
-      ) : null}
-      {!markdown && !office ? (
-        <dl>
-          <div><dt>{t("details.id")}</dt><dd>{artifact.id}</dd></div>
-          {detail?.mimeType || artifact.mimeType ? <div><dt>{t("details.type")}</dt><dd>{detail?.mimeType || artifact.mimeType}</dd></div> : null}
-        </dl>
-      ) : null}
-      {loading ? <p aria-live="polite">{t("details.loadingArtifact")}</p> : null}
-      {error ? <p role="alert">{error}</p> : null}
-      {notice ? <p className="react-artifact-detail__notice">{notice}</p> : null}
-      {detail?.imageDataUrl ? <img alt={detail.title} src={detail.imageDataUrl} /> : null}
-      {detail?.dataView ? <DataViewCard artifact={{ ...artifact, dataView: detail.dataView }} expanded /> : null}
-      {office ? (
-        <OfficeArtifactPreview
-          onAskForContentChange={!error && localThreadId && file.revision && artifact.fetchPath ? (request) => {
-            const position = request.start === request.end ? String(request.start) : `${request.start}–${request.end}`;
-            onReference(officeContentReference({ request, path: artifact.fetchPath!, title: artifact.title, threadId: localThreadId, revision: file.revision!,
-              label: t(request.kind === "document" ? "details.officeParagraphSelection" : "details.officeSlideSelection", { position }),
-            }));
-          } : undefined}
-          onAskForChange={error ? undefined : (selection) => onAskForSpreadsheetChange(artifact, selection, file.revision)}
-          source={office}
-        />
-      ) : null}
-      {markdownContent ? (
-        <article aria-label={markdownContent.title} className="react-artifact-detail__document" role="document">
-          <AssistantMarkdown
-            onOpenFileLink={onOpenFileLink}
-            streaming={false}
-            text={markdownContent.text}
-          />
-        </article>
-      ) : detail?.textContent ? <pre className="react-artifact-detail__text">{detail.textContent}</pre> : null}
-      {!loading && !error && !office && !detail?.dataView && !detail?.imageDataUrl && !detail?.textContent ? <p>{t("details.noPreview")}</p> : null}
-    </div>
-  );
-}
-
-function isMarkdownArtifact(artifact: ArtifactRef, detail?: LoadedArtifactDetail): boolean {
-  const mimeType = (detail?.mimeType || artifact.mimeType || "").split(";", 1)[0].trim().toLowerCase();
-  return artifact.kind.toLowerCase() === "markdown" || mimeType === "text/markdown";
-}
-
 function toolCallDetailSections(toolCall: ToolCallSummary, t: TFunction<"chat">): Array<{ label: string; value: string }> {
   return [
     { label: t("details.status"), value: toolCall.status },
@@ -2362,19 +1790,6 @@ function formatDetailLines(rows: Array<[string, string | undefined]>): string {
     .join("\n");
 }
 
-function browserResourceTitle(title: string, url: string, fallback: string): string {
-  const normalizedTitle = title.trim();
-  if (normalizedTitle && normalizedTitle !== "about:blank" && normalizedTitle !== "New tab") {
-    return normalizedTitle;
-  }
-  if (!url || url === "about:blank") return fallback;
-  try {
-    return new URL(url).hostname || url;
-  } catch {
-    return url;
-  }
-}
-
 function projectDraftSessionSummary(draft: DraftSession): SessionSummary {
   return {
     id: draft.id,
@@ -2387,17 +1802,7 @@ function projectDraftSessionSummary(draft: DraftSession): SessionSummary {
   };
 }
 
-function omitRecordKey<T>(record: Record<string, T>, key: string): Record<string, T> {
-  if (!(key in record)) return record;
-  const next = { ...record };
-  delete next[key];
-  return next;
-}
-
 function boundedSpreadsheetSelectionValue(value: string): string {
   return value.length > 12000 ? `${value.slice(0, 12000)}\n[Selection excerpt truncated; read the referenced range for all values.]` : value;
 }
 
-function errorMessage(error: unknown): string {
-  return error instanceof Error ? error.message : String(error);
-}
