+++ src/react-workbench/chat/useChatSessionRuntime.test.tsx
@@ -4,7 +4,6 @@ import { act, renderHook, waitFor } from "@testing-library/react";
 import { afterEach, describe, expect, test, vi } from "vitest";
 import type { AgentUiForm } from "../../app-core/agent-ui/agentUiEvents";
 import type { ChatTimelineSnapshot } from "../../app-core/chat/agentTimelineModel";
-import { createNativeBrowserSessionSnapshot } from "../../app-core/native/nativeBrowserSnapshot";
 import type { ChatEvent } from "../services";
 import {
   useChatSessionRuntime,
@@ -127,34 +126,6 @@ describe("useChatSessionRuntime", () => {
     expect(result.current.state).toMatchObject({ error: "forms unavailable", status: "failed" });
   });
 
-  test("does not replace a newer visible Browser snapshot with a stale hidden response", async () => {
-    let listener: ((event: ChatEvent) => void) | undefined;
-    const store = runtimeStore({
-      subscribe: vi.fn((_sessionId, nextListener) => {
-        listener = nextListener;
-        return vi.fn();
-      }),
-    });
-    const { result } = renderHook(() => useChatSessionRuntime({
-      chatStore: store,
-      sessionId: "session-1",
-    }));
-    await waitFor(() => expect(result.current.state.status).toBe("ready"));
-
-    act(() => listener?.({
-      browserSnapshot: nativeBrowserSnapshot(3, "visible"),
-      type: "browser.snapshot",
-    }));
-    await waitFor(() => expect(result.current.state.browserSnapshot?.data.surface?.lifecycle).toBe("visible"));
-    act(() => listener?.({
-      browserSnapshot: nativeBrowserSnapshot(2, "hidden"),
-      type: "browser.snapshot",
-    }));
-
-    expect(result.current.state.browserSnapshot?.data.surface?.lifecycle).toBe("visible");
-    expect(result.current.state.browserSnapshot?.revision).toBe(3);
-  });
-
   test("keeps a live hook result when the initial timeline finishes loading later", async () => {
     let listener: ((event: ChatEvent) => void) | undefined;
     let resolveTimeline: ((value: ChatTimelineSnapshot) => void) | undefined;
@@ -221,32 +192,3 @@ function runningTimeline(sessionId: string, turnId: string): ChatTimelineSnapsho
   };
 }
 
-function nativeBrowserSnapshot(revision: number, surfaceLifecycle: "hidden" | "visible") {
-  return createNativeBrowserSessionSnapshot({
-    activeTabId: "browser-tab-1",
-    browserSessionId: "browser-session-1",
-    contract: "browser_session_v1",
-    interaction: { click: true, navigate: true, type: true },
-    kind: "browser_session",
-    lifecycle: "ready",
-    operationId: "browser-operation-1",
-    runtimeKind: "windows_webview2",
-    sessionId: "session-1",
-    state: "running",
-    surface: { layoutRevision: revision, lifecycle: surfaceLifecycle },
-    tabs: [{
-      activeHistoryIndex: 0,
-      captures: [],
-      history: [{ title: "Example", url: "https://example.com" }],
-      loading: false,
-      rendererLifecycle: "running",
-      tabId: "browser-tab-1",
-      title: "Example",
-      url: "https://example.com",
-    }],
-  }, {
-    observedAt: "2026-08-18T08:00:00Z",
-    revision,
-    sourceId: "native-browser:browser-session-1",
-  });
-}
